front: move NGE logic to dedicated component directory
Move low-level NGE details to a separate directory to make sure it
stays clearly split from the OSRD-specific logic.
