+++ front/src/applications/operationalStudies/components/NGE/types.ts
